requirements.txt
praw==7.5.0
requests==2.27.1
datetime==4.3
google-api-python-client==2.37.0 
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
